test-automation/sanity_check/Testcases/requirements.txt
appdirs==1.4.4
Appium-Python-Client==2.11.1
apturl==0.5.2
async-generator==1.10
attrs==23.1.0
Automat==20.2.0
Babel==2.8.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.12.2
beniget==0.4.1
blinker==1.6.2
Brlapi==0.8.3
Brotli==1.0.9
certifi==2022.12.7
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.2.0
chromedriver-autoinstaller==0.6.2
click==8.1.7
colorama==0.4.4
command-not-found==0.3
constantly==15.1.0
cryptography==3.4.8
cssselect==1.1.0
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
decorator==4.4.2
defer==1.0.6
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.1
duplicity==0.8.21
exceptiongroup==1.1.1
fasteners==0.14.1
filelock==3.6.0
Flask==2.3.3
fonttools==4.29.1
fs==2.4.12
future==0.18.2
gast==0.5.2
google==3.0.0
greenlet==2.0.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
html5lib==1.1
httplib2==0.20.2
hyperframe==6.0.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
iniconfig==2.0.0
ipython==7.31.1
itemadapter==0.4.0
itemloaders==1.0.4
itsdangerous==2.1.2
jedi==0.18.0
jeepney==0.7.1
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.3.2
jsonschema==4.17.3
keyboard==0.13.5
keyring==23.5.0
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.3
matplotlib==3.5.1
matplotlib-inline==0.1.3
monotonic==1.6
more-itertools==8.10.0
mpmath==0.0.0
mysql-connector==2.2.9
mysql-connector-python==8.1.0
netifaces==0.11.0
networkx==3.1
nltk==3.8.1
numpy==1.24.3
oauthlib==3.2.0
olefile==0.46
outcome==1.2.0
packaging==23.1
pandas==2.0.2
paramiko==2.9.3
parsel==1.6.0
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.5.1
pluggy==1.0.0
ply==3.11
prompt-toolkit==3.0.28
Protego==0.2.1
protobuf==4.21.12
psycopg2-binary==2.9.7
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
PyDispatcher==2.0.5
pydub==0.25.1
Pygments==2.11.2
PyGObject==3.42.1
PyHamcrest==2.0.2
pyinotify==0.9.6
PyJWT==2.3.0
pylinkvalidator==0.3
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.19.3
PySocks==1.7.1
pytest==7.3.2
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.0
python-owasp-zap-v2.4==0.0.21
pythran==0.10.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzap==0.3.0
queuelib==1.6.2
regex==2023.8.8
reportlab==3.6.8
requests==2.31.0
scikit-learn==1.3.0
scipy==1.11.1
Scrapy==2.5.1
SecretStorage==3.3.1
selenium==4.11.2
service-identity==18.1.0
simplejson==3.17.6
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SpeechRecognition==3.10.0
SQLAlchemy==2.0.21
sympy==1.9
systemd-python==234
theHarvester==0.0.1
threadpoolctl==3.2.0
toml==0.10.2
tomli==2.0.1
tqdm==4.65.0
traitlets==5.1.1
trio==0.22.0
trio-websocket==0.10.2
Twisted==22.1.0
typing_extensions==4.8.0
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
urllib3==1.26.5
usb-creator==0.3.7
vboxapi==1.0
virtualenv==20.13.0+ds
w3lib==1.22.0
wadllib==1.3.6
wcwidth==0.2.5
webdriver-manager==4.0.0
webencodings==0.5.1
Werkzeug==2.3.7
wsproto==1.2.0
xdg==5
xkit==0.0.0
zipp==1.0.0
zope.interface==5.4.0